DOCS: DTO CQRS 패턴에 맞게 위치 변경
