+++ src/main/java/spot/spot/domain/job/command/dto/request/RegisterWorkerRequest.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.job.dto.request;
+package spot.spot.domain.job.command.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
